event handling to ship class
